requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
numpy>=1.20.0
PyQt5>=5.15.4
pygame>=2.0.1
pillow>=8.3.1
mediapipe>=0.10.0
scipy>=1.7.0